Modify args from hook OnExplosiveThrown and OnExplosiveDropped
Adding the third argument to the OnExplosiveThrown and OnExplosiveDropped hooks that will be carried by ThrownWeapon. 
This is done in order to be able to know from which of the entities the BaseEntity was spawned which is passed by the second argument. The work of old plug-ins does not affect because they will ignore the third argument.