Add new Hooks and Modifiers
Hooks:

CanLootEntity for BaseRidableAnimal - CanLootEntity extension, now it works on horses.
CanSamSiteTargeted - It is executed at the moment when SamSite is looking for a target. The logical replacement for the OnSamSiteTarget hook that was created when the OxidePatcher was broken.
OnCloseConnection - It is executed when Connection with the player is closed.

Modifiers:
ResourceEntity::resourceDispenser - Allows you to edit the dispenser of extracted resources.
BaseCorpse::resourceDispenser - Allows you to edit the corpse dispenser.