Fix hooks for selecting vehicle modules
Removed Item argument to OnVehicleModuleDeselected because the RPC message doesn't seem to contain the information required to get it anyway.

Changed OnVehicleModuleSelected to pass an Item instead of a BaseVehicleModule to be consistent with the OnVehicleModuleSelect hook. Also moved the hook call to a different line so that it's called for all module types instead of just for storage modules.