Added Hook "CanDropBackpack"
The hook prevents the backpack from falling out in the Wounded, Crawling, or Dead states.

Upon death, it will still drop if the backpack is not empty, due to the CanAcceptItem check. By changing containerVolumeWhenFilled to zero for ItemModBackpack, you can allow the backpack to move through containers, even if it is not empty.

The hook is also needed to access the backpack when dying, as the backpack will fall out before the OnPlayerDeath hook is triggered.